requirements.txt
flask
torch
torchvision
pillow
scikit-learn
matplotlib
opencv-python
